+++ src/core/scenario.jl
@@ -28,6 +28,7 @@
 # OF THE POSSIBILITY OF SUCH DAMAGE.
 # *********************************************************************************
 struct Scenario
+    settings::Settings
     site::Site
     pvs::Array{PV, 1}
     storage::Storage
@@ -66,7 +67,11 @@ end
 ```
 """
 function Scenario(d::Dict)
-    # TODO add validate! functions for each input struct
+    if haskey(d, "Settings")
+        settings = Settings(;dictkeys_tosymbols(d["Settings"])...)
+    else
+        settings = Settings()
+    end
     
     site = Site(;dictkeys_tosymbols(d["Site"])...)
     
@@ -109,7 +114,9 @@ function Scenario(d::Dict)
         storage = Storage(storage_dict, financial)
     end
 
-    electric_load = ElectricLoad(; dictkeys_tosymbols(d["ElectricLoad"])...)
+    electric_load = ElectricLoad(; dictkeys_tosymbols(d["ElectricLoad"])...,
+                                   latitude=site.latitude, longitude=site.longitude
+                                )
 
     electric_tariff = ElectricTariff(; dictkeys_tosymbols(d["ElectricTariff"])..., 
                                        year=electric_load.year
@@ -122,6 +129,7 @@ function Scenario(d::Dict)
     end
 
     return Scenario(
+        settings,
         site, 
         pvs, 
         storage,